Rename machine to dfa.
