+++ revex/dfa.py
